# api/openapi.yaml
openapi: 3.0.3
info:
  title: BlitterServer API
  version: 1.0.0-draft.3
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: |
    The BlitterAmp middleman service. Apps (BlitterAmp iOS/desktop) talk only to
    this API; BlitterServer owns the music source (Plex in v1), server-side
    transcoding, acquisition (Lidarr), last.fm, and cross-device taste/love
    state.

    **Auth realms:** device endpoints use bearer tokens; `/admin/api/*` uses an
    admin cookie session. Server push is SSE — see `GET /v1/events` for the
    envelope and event catalog.

    **Users are household profiles** (Netflix-style, passwordless, optional
    PIN; created by the admin in the web admin UI). PIN pairing mints a
    **device token** whose only powers are listing profiles and exchanging for
    **profile tokens** (`POST /v1/profile-tokens`). Everything else requires a
    profile token and is implicitly scoped to that profile: playlists,
    ratings, taste, loves, play history, home rails, last.fm connection.
    Shared across profiles: the library, the artifact cache, and acquisition
    (new-release watching is the union of everyone's artist loves).
    Profile switching in-app = using a different cached profile token; no
    admin involvement after the one-time device pairing approval.

    **Conventions:** JSON everywhere; list endpoints paginate
    with `cursor`/`limit` and return `{items, nextCursor}`; errors are a Problem object.

...
servers:
...
security:
  - deviceToken: []
paths:
...
  /v1/pair:
    post:
      operationId: startPairing
      tags: [pairing]
      summary: Begin PIN pairing
      description: |
        Returns a short-lived human-readable code the user approves in the web
        admin. The app then polls `GET /v1/pair/{pairingId}`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName, deviceType]
              properties:
                deviceName: { type: string, example: "Nathan's iPhone" }
                deviceType: { type: string, enum: [ios, desktop, web, other] }
                appVersion: { type: string }
      responses:
        "201":
          description: Pairing started
          content:
            application/json:
              schema:
                type: object
                required: [pairingId, code, expiresAt]
                properties:
                  pairingId: { type: string }
                  code: { type: string, example: "7GK4PC", description: Show to the user; entered/approved in web admin }
                  expiresAt: { type: string, format: date-time }
        "429": { $ref: "#/components/responses/Problem" }
  /v1/pair/{pairingId}:
    get:
      operationId: getPairing
      tags: [pairing]
      summary: Poll pairing state
      security: []
      parameters:
        - { name: pairingId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Pairing state
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [pending, approved, denied, expired] }
                  token: { type: string, nullable: true, description: "DEVICE token, present once when status=approved — exchange it for a profile token via /v1/profile-tokens" }
                  deviceId: { type: string, nullable: true }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/pair/claim:
    post:
      operationId: claimPairCode
      tags: [pairing]
      summary: Claim a QR pairing code (one scan = discovery + auth)
      description: |
        The QR shown in the web admin encodes the server's canonical URL plus a
        short-lived (≤15 min) single-use code. Claiming mints a **device
        token** immediately — no admin approval step (showing the QR was the
        approval). The app then lists profiles and mints profile tokens as
        usual.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, deviceName, deviceType]
              properties:
                code: { type: string, example: "7GK4PC" }
                deviceName: { type: string, example: "Nathan's iPhone" }
                deviceType: { type: string, enum: [ios, desktop, web, other] }
                appVersion: { type: string }
      responses:
        "201":
          description: Device paired
          content:
            application/json:
              schema:
                type: object
                required: [token, deviceId]
                properties:
                  token: { type: string, description: Device token — exchange via /v1/profile-tokens }
                  deviceId: { type: string, example: dev_8f2k1q }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { description: Code expired or already used, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "429": { $ref: "#/components/responses/Problem" }

  # ───────────────────────────── profiles ─────────────────────────────
...
  /v1/stream/{trackId}:
    get:
      operationId: streamTrack
      tags: [streaming]
      summary: Stream the original file (direct play, Range supported)
      parameters:
        - { name: trackId, in: path, required: true, schema: { type: string } }
        - { name: Range, in: header, required: false, schema: { type: string } }
      responses:
        "200": { description: Full content, content: { audio/*: { schema: { type: string, format: binary } } } }
        "206": { description: Partial content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/stream-grants:
    post:
      operationId: createStreamGrant
      tags: [streaming]
      summary: Mint a short-lived signed URL for players that cannot send headers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackId]
              properties:
                trackId: { type: string }
      responses:
        "201":
          description: Grant
          content:
            application/json:
              schema:
                type: object
                required: [url, expiresAt]
                properties:
                  url: { type: string, description: "Absolute URL, valid for the grant lifetime only" }
                  expiresAt: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/art/{artId}:
    get:
      operationId: getArt
      tags: [streaming]
      summary: Artwork with server-side resize + cache
      parameters:
        - { name: artId, in: path, required: true, schema: { type: string } }
        - { name: w, in: query, required: false, schema: { type: integer, maximum: 2048 } }
        - { name: h, in: query, required: false, schema: { type: integer, maximum: 2048 } }
      responses:
        "200": { description: Image, content: { image/jpeg: { schema: { type: string, format: binary } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ───────────────────────────── artifacts ─────────────────────────────
  /v1/artifacts:
    post:
      operationId: requestArtifacts
      tags: [artifacts]
      summary: Request download artifacts (server transcodes + caches; batch-friendly)
      description: |
        Idempotent per (trackId, format, bitrateKbps): re-requesting returns the
        existing artifact. `format: original` artifacts are immediately `ready`.
        Progress and readiness arrive via the `artifact.updated` event on
        `GET /v1/events`.

        Each artifact carries `albumId` + `artId` so sync flows can prefetch
        album art (one fetch per distinct albumId) and downloaded content
        renders with art offline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackIds, format]
              properties:
                trackIds:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { type: string }
                format: { type: string, enum: [original, aac] }
                bitrateKbps: { type: integer, enum: [128, 192, 256, 320], nullable: true, description: Required for aac }
      responses:
        "202":
          description: Artifacts (existing or queued)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Artifact" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/artifacts/{artifactId}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: Artifact status
      parameters: [{ name: artifactId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Artifact
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: releaseArtifact
      tags: [artifacts]
      summary: Hint that this device no longer needs the artifact (server may evict)
      parameters: [{ name: artifactId, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Released }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/events:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Server-sent events — typed event stream
      description: |
        SSE (`text/event-stream`), authenticated with a profile token. Each
        message's `id:` field is a monotonic sequence number; reconnect with
        `Last-Event-ID` to resume without loss. `data:` is an Event envelope
        (`{type, at, data}`). Clients MUST ignore unknown types.

        | type | data | scope |
        |---|---|---|
        | `library.changed` | `{libraryId, updatedAt}` | instance |
        | `artifact.updated` | `Artifact` | instance |
        | `acquisition.updated` | `AcquisitionActivity` | instance |
        | `presence.updated` | `PresenceEntry` | instance (honors share-listening) |
        | `love.updated` | `LoveRecord` | profile |
        | `playlist.changed` | `{playlistId}` | profile (incl. shared/collab changes by others) |
        | `taste.updated` | `{}` | profile |
        | `recommendation.received` | `Recommendation` | profile |
        | `party.invited` | `PartyInvite` | profile |
        | `party.updated` | `PartyState` (with `partyId`) | party members |

        There is no client→server channel — control actions are ordinary
        POSTs. Every event-producing resource keeps a plain pollable GET as
        the resync path (mandatory after iOS backgrounding kills the stream).

        There is no presence-removal event — a profile stopping playback or
        disabling share-listening simply stops appearing in `GET /v1/presence`;
        clients converge via that poll path.
      parameters:
        - { name: Last-Event-ID, in: header, required: false, schema: { type: string }, description: Sequence number of the last event received }
      responses:
        "200":
          description: Event stream (unbounded)
          content:
            text/event-stream:
              schema: { type: string, description: SSE frames whose data lines carry Event JSON }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ───────────────────────────── admin ─────────────────────────────
...
  /admin/api/session:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login (sets session cookie)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "204": { description: Session cookie set }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      security: [{ adminSession: [] }]
      responses:
        "204": { description: Logged out }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: Per-device token minted via PIN pairing
    adminSession:
      type: apiKey
      in: cookie
      name: blitter_admin
...
  responses:
    Unauthorized:
      description: Missing/invalid credentials
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string, nullable: true }
        code: { type: string, nullable: true, description: "Machine-readable, e.g. source_unavailable, artifact_not_ready" }
...
    Artifact:
      type: object
      required: [artifactId, trackId, format, status]
      properties:
        artifactId: { type: string, example: arf_3e4r5t }
        trackId: { type: string, example: trk_2vq8na }
        albumId: { type: string, nullable: true, example: alb_7xm1pd, description: "Grouping key for offline art (one art fetch per album, not per track)" }
        artId: { type: string, nullable: true, example: img_4c5d6e, description: "Album art for the track — sync flows prefetch /v1/art/{artId} alongside the file so downloaded content has art offline" }
        format: { type: string, enum: [original, aac] }
        bitrateKbps: { type: integer, nullable: true }
        status: { type: string, enum: [queued, processing, ready, failed] }
        bytes: { type: integer, format: int64, nullable: true, description: Exact size once ready }
        error: { type: string, nullable: true }
...